Change sensitive user properties in authentication context
